+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-v0.2.2
+v0.3.0
 ======
 
 * Added Prim's Algorithm (`alg.prim`).